refactor folders, added crosshair(no func)
