+++ source/main.lua
@@ -16,5 +16,5 @@ Noble.GameData.setup({
 Noble.showFPS = true
 playdate.display.setRefreshRate(60)
 
-Noble.new(InitialScene, 1.5, Noble.TransitionType.CROSS_DISSOLVE)
--- Noble.new(SpaceScene, 1.5, Noble.TransitionType.CROSS_DISSOLVE)
+-- Noble.new(InitialScene, 1.5, Noble.TransitionType.CROSS_DISSOLVE)
+Noble.new(SpaceScene, 1.5, Noble.TransitionType.CROSS_DISSOLVE)
